+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # developing
 - [Fix 4984: random expire cache](https://github.com/beego/beego/pull/4984)
-- [Feat 4999: add get all tasks function #](https://github.com/beego/beego/pull/4999)
 - [Fix 4907: make admin serve HTTP only](https://github.com/beego/beego/pull/5005)
+- [Feat 4999: add get all tasks function](https://github.com/beego/beego/pull/4999)
 
 
 # v2.0.4