+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [Fix 4907: make admin serve HTTP only](https://github.com/beego/beego/pull/5005)
 - [Feat 4999: add get all tasks function](https://github.com/beego/beego/pull/4999)
 - [Fix 5012: fix some bug, pass []any as any in variadic function](https://github.com/beego/beego/pull/5012)
+- [Fix 5022: Miss assigning listener to graceful Server](https://github.com/beego/beego/pull/5028)
 
 
 # v2.0.4