+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # developing
+
+# v2.0.5
+
+Note: now we force the web admin service serving HTTP only.
+
 - [Fix 4984: random expire cache](https://github.com/beego/beego/pull/4984)
 - [Fix 4907: make admin serve HTTP only](https://github.com/beego/beego/pull/5005)
 - [Feat 4999: add get all tasks function](https://github.com/beego/beego/pull/4999)